crewai-outbound/multiagents/requirements.txt
crewai>=0.51.0
crewai-tools>=0.4.0
python-dotenv>=1.0.1
requests>=2.32.3
urllib3>=2.2.2
PyYAML>=6.0.1
pydantic>=2.7.0
slack_sdk>=3.30.0
